# Fixify Solutions/Fixify-NodeJS-BackEnd/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Fixify Backend API
  version: 1.0.0
  description: Backend API for Fixify (auth, users, services)
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /logs/backend-logs/{filename}:
    get:
      summary: Read full content of a log file
      tags: [Logs]
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log file content
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  content: { type: string }
  /logs/backend-logs/{filename}/download:
    get:
      summary: Download entire log file
      tags: [Logs]
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /logs/backend-logs/{filename}/last/{count}:
    get:
      summary: Get last N lines from a log file
      tags: [Logs]
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
        - in: path
          name: count
          required: true
          schema:
            type: integer
            example: 50
      responses:
        "200":
          description: Last N lines of the log file
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  count: { type: integer }
                  content: { type: string }
  /logs/backend-logs/{filename}/last/{count}/download:
    get:
      summary: Download last N lines from a log file
      tags: [Logs]
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
        - in: path
          name: count
          required: true
          schema:
            type: integer
            example: 100
      responses:
        "200":
          description: Partial log download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary